refractor moved functions to utils/functions file
