+++ src/contracts/functions.ts
@@ -12,6 +12,113 @@ const CONTRACT_ADDR = "0x7f60e977a7b9677be1795efe5ad5516866ab69a6";
 
 type StampType = number;
 
+export async function fetchDiscount(
+  aikoAPI: string,
+  existsId: number,
+  stampsHeld: number,
+  key: string
+): Promise<string> {
+  try {
+    const response = await fetch(`${aikoAPI}/get-discount/${existsId}`);
+    if (!response.ok) {
+      throw new Error(`HTTP error! status: ${response.status}`);
+    }
+    const data = await response.json();
+    const object = {
+      rewards: stampsHeld >= 9 ? 3 : stampsHeld >= 6 ? 2 : 1,
+      wallet: data.discount_codes[0].code,
+    };
+    return encrypt(object, key);
+  } catch (error) {
+    console.error("Error:", error);
+    return "";
+  }
+}
+
+export async function postDiscount(
+  aikoAPI: string,
+  walletAddress: string,
+  stampsHeld: number,
+  key: string
+): Promise<string> {
+  try {
+    const response = await fetch(`${aikoAPI}/get-discounts`);
+    if (!response.ok) {
+      throw new Error(`HTTP error! status: ${response.status}`);
+    }
+
+    const data = await response.json();
+    const discounts = data.price_rules;
+
+    const exists = discounts.find(
+      (discount: any) => discount.title === walletAddress
+    );
+
+    if (exists) {
+      return fetchDiscount(aikoAPI, exists.id, stampsHeld, key);
+    }
+
+    const couponCode = generateRandomCouponCode(10);
+    const object = {
+      rewards: stampsHeld >= 9 ? 3 : stampsHeld >= 6 ? 2 : 1,
+      wallet: couponCode,
+    };
+    const encryptedObject = encrypt(object, key);
+
+    let products: string[] = [];
+
+    if (stampsHeld >= 9) {
+      products = [
+        "gid://shopify/Product/8312585650467",
+        "gid://shopify/Product/8312587583779",
+        "gid://shopify/Product/8312588566819",
+      ];
+    } else if (stampsHeld >= 6) {
+      products = [
+        "gid://shopify/Product/8312585650467",
+        "gid://shopify/Product/8312587583779",
+      ];
+    } else if (stampsHeld >= 3) {
+      products = ["gid://shopify/Product/8312585650467"];
+    }
+
+    const basicCodeDiscount = {
+      title: `${walletAddress}`,
+      code: `${couponCode}`,
+      startsAt: "2022-06-21T00:00:00Z",
+      endsAt: "2023-09-21T00:00:00Z",
+      customerSelection: { all: true },
+      customerGets: {
+        value: { percentage: 1 },
+        items: {
+          products: {
+            productsToAdd: products,
+          },
+        },
+      },
+      appliesOncePerCustomer: false,
+      usageLimit: 1,
+    };
+
+    const postResponse = await fetch(`${aikoAPI}/generate-discount`, {
+      method: "POST",
+      headers: {
+        "Content-Type": "application/json",
+      },
+      body: JSON.stringify({ basicCodeDiscount }),
+    });
+
+    if (!postResponse.ok) {
+      throw new Error(`HTTP error! status: ${postResponse.status}`);
+    }
+
+    return encryptedObject;
+  } catch (error) {
+    console.error("Error:", error);
+    throw error;
+  }
+}
+
 export function toBase64Url(base64: string): string {
   return base64.replace("+", "-").replace("/", "_").replace(/=+$/, "");
 }
